Refactor logic to untoggle into custom hook
