+++ src/hooks/useUntoggle.js
@@ -1,0 +1,23 @@
+import { useEffect, useState, useRef } from "react";
+
+const useUntoggle = () => {
+  const [expanded, setExpanded] = useState(false);
+  const ref = useRef(null);
+
+  useEffect(() => {
+    const handleClickOutside = (event) => {
+      if (ref.current && !ref.current.contains(event.target)) {
+        setExpanded(false);
+      }
+    };
+
+    document.addEventListener("mouseup", handleClickOutside);
+    return () => {
+      document.removeEventListener("mouseup", handleClickOutside);
+    };
+  }, [ref]);
+
+  return { expanded, setExpanded, ref };
+};
+
+export default useUntoggle;
